refactor components into separate files/directories
