+++ src/App.css
@@ -1,3 +1,32 @@
+/* navbar styling */
+
+ul {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+  overflow: hidden;
+  background-color: #333;
+}
+
+li {
+  float: left;
+}
+
+li a {
+  display: block;
+  color: white;
+  text-align: center;
+  padding: 14px 16px;
+  text-decoration: none;
+}
+
+/* Change the navbar link color to #111 (black) on hover */
+li a:hover {
+  background-color: #111;
+}
+
+/* table styling */
+
 #org-table {
   font-family: Arial, Helvetica, sans-serif;
   border-collapse: collapse;
