+++ src/explore/searchBar.jsx
@@ -1,0 +1,21 @@
+import React, { Component } from 'react';
+
+class SearchBar extends Component {
+  constructor(props) {
+    super(props);
+    this.handleFilterTextInputChange = this.handleFilterTextInputChange.bind(this);
+  }
+  handleFilterTextInputChange(e) {
+    this.props.onFilterTextInput(e.target.value);
+  }
+  render () {
+    return (
+      <form>
+        <div>Search for stuff here. </div>
+        <input type="text" placeholder="Search..." value={this.props.filterText} onChange={this.handleFilterTextInputChange}/>
+      </form>
+    );
+  }
+}
+
+export default SearchBar;
