+++ src/index.css
@@ -1,3 +1,4 @@
+/*global styles here*/
 body {
   margin: 0;
   padding: 0;
