+++ src/index.js
@@ -1,7 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-// import WholeApp from './App';
 import Navigation from './App';
 import registerServiceWorker from './registerServiceWorker';
