+++ src/info.jsx
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+const Info = () => (
+  <div className='volunteer-info'>
+    <p>Info to go here</p>
+  </div>
+)
+
+export default Info;
