+++ src/map.jsx
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+const Map = () => (
+  <div>
+    <p>Map to go here</p>
+  </div>
+)
+
+export default Map;
